+++ gradle.properties
@@ -15,7 +15,7 @@ loader_version_range=[4,)
 mod_id=journeymode
 mod_name=Journey Mode
 mod_license=MIT
-mod_version=1.4.2N
+mod_version=1.5.0N
 mod_group_id=com.aryangpt007.journeymode
 mod_authors=Aryangpt007
 mod_description=Unlock unlimited access to items after collecting enough of them
